Use the older container cache
To better mimic the juno deployments and to be more
compatible with the base Ubuntu Trusty (14.04.0)
installation being used, use the old container cache.

In the AIO bootstrap the cache is adjusted to ensure
that it doesn't get any updated packages and that it
is using the base packages that matter.